Adds toResult function analogous to toMaybe
toResult is somewhat compatible with the idea of
toMaybe in that it different constructur per Success then for other
cases.

For api it means that function must take is parametrized so that
it requires default error from the caller.

This means the implementation is equivalent to composition of
Result.fromMaybe and toMaybe in following way:

    toResult e == Maybe.fromResult e << toMaybe

But is not implemented using explicit pattern matching for
implementation clarity and performance reasons.